requirements.txt
python-telegram-bot>=20.0
anthropic>=0.40.0
google-cloud-speech>=2.0.0
google-cloud-texttospeech>=2.0.0
pydub>=0.25.1
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
